libdiskfs: Port to x86_64
Message-Id: <20230508213136.608575-9-bugaevc@gmail.com>

+++ libdiskfs/sync-interval.c
@@ -91,9 +91,9 @@ diskfs_set_sync_interval (int interval)
    INTERVAL seconds, as long as it's in the thread pointed to by the global
    variable PERIODIC_SYNC_THREAD.   */
 static void *
-periodic_sync (void * arg)
+periodic_sync (void *arg)
 {
-  int interval = (int) arg;
+  int interval = (int)(uintptr_t) arg;
   for (;;)
     {
       error_t err;
